third_party/blink/web_tests/external/wpt/css/css-fonts/font-face-style-normal-expected.txt

This is a testharness.js-based test.
[FAIL] CSS Fonts: parsing 'normal' in the font-style descriptor 1
  assert_equals: 'oblique 0deg' serializes as 'normal' expected "normal" but got "oblique 0deg"
Harness: the test ran to completion.
